Make most of patient search optional
The main use case is for patients to access their own information. Valid and reasonable implementations may not want to support searching for other patients.

There certainly is a use for acting on behalf of a relative, for instance, and an app might want to sort patients whose information it has access to based on some criteria. But still, I'd mark that as recommended, not mandatory.

We should also think of search parameters for other resources. Do we usually need the patient to be  parameter? I would guess no, but since there is the scenario described above, it's OK for me to leave those as they are.