Update authors
Added Eric. Updated sorting to be alphabetical (by last name). That's what makes sense, right?

Or do we want to have the authors listed here at all? We have the contributors on the GitHub page, and the [credits](https://www.hl7.org/fhir/uv/ipa/#credits) section in the spec...